% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmc.center.heatmap.R
\name{fcmc.center.heatmap}
\alias{fcmc.center.heatmap}
\title{Drow heatmap of the correlation of centers about the fcmc result}
\usage{
fcmc.center.heatmap(data)
}
\arguments{
\item{data}{the fcmc$centers matrix}
}
\value{

}
\description{
Drow heatmap of the correlation of centers about the fcmc result
}
\examples{
a ggfcmc result
fcmc.center.heatmap(t(fcmc$centers))
}
