% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmc.df.R
\name{fcmc.df}
\alias{fcmc.df}
\title{Get the cluster and membership information}
\usage{
fcmc.df(data.mean, fcmc, write = TRUE)
}
\arguments{
\item{data.mean}{a matrix from grp.mean function}

\item{fcmc}{the result of ggfcmc function}

\item{write}{a boolean value, save the result or not}
}
\value{
a data frame
}
\description{
Get the cluster and membership information
}
\examples{
fcmc.df(data.mean = genus.syn.mean, fcmc = fcmc, write = F)
}
\author{
Zhonghui Gai
}
