% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmcPCA.R
\name{fcmcPCA}
\alias{fcmcPCA}
\title{PCA plot based on ggplot2 using the stats::prcomp function}
\usage{
fcmcPCA(mean.data, group, pc = 12, level = 0.68)
}
\arguments{
\item{mean.data}{A data frame containing the first collumn as the grouping information}

\item{group}{the result of fcmc, default value is fcmc$cluster}

\item{pc}{Which PC to present, one of 12, 13, 23, the default value is 12}

\item{level}{default value is 0.68}
}
\value{

}
\description{
PCA plot based on ggplot2 using the stats::prcomp function
}
\examples{
fcmcPCA(genus.syn.mean, fcmc$cluster, pc = 12, level = 0.95) +
 ggsci::scale_color_aaas()
}
\author{
ZhonghuiGai
}
