% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.std.R
\name{filter.std}
\alias{filter.std}
\title{Filter data step 1}
\usage{
filter.std(data, min.std = 0)
}
\arguments{
\item{data}{dataframe}

\item{min.std}{min standard deviation}
}
\value{
a filted data
}
\description{
Filter data step 1
}
\author{
Zhonghui Gai
}
