% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcenterplot.R
\name{ggcenterplot}
\alias{ggcenterplot}
\title{Draw the centroid profiles step 5}
\usage{
ggcenterplot(data)
}
\arguments{
\item{data}{result from ggcfmc or cmeans}
}
\value{

}
\description{
Draw the centroid profiles step 5
}
\examples{
ggcenterplot(data = fcmc)
}
\author{
Zhonghui Gai
}
