% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggclusterplot.R
\name{ggclusterplot}
\alias{ggclusterplot}
\title{Draw the cluster plots}
\usage{
ggclusterplot(data, fcm)
}
\arguments{
\item{data}{scaled original mean data created by grp.mean function.}

\item{fcm}{a aojedct created from ggfcmc or cmeans functions.}
}
\value{
cluster photoes
}
\description{
Draw the cluster plots
}
\author{
Zhonghui Gai
}
