% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfcmc.R
\name{ggfcmc}
\alias{ggfcmc}
\title{Function for soft clustering based on fuzzy c-means.}
\usage{
ggfcmc(df, centers, m, ...)
}
\arguments{
\item{df}{The data matrix where columns correspond to variables
and rows to observations.}

\item{centers}{Number of clusters or initial values for cluster centers.}

\item{m}{A number greater than 1 giving the degree of fuzzification.}

\item{...}{other parameter for cmeans in e1071 pkg.}
}
\value{

}
\description{
Function for soft clustering based on fuzzy c-means.
}
\examples{
fcmc <- ggfcmc(genus.syn.mean, centers = 8, m = mestimate(genus.syn.mean))
}
\author{
Zhonghui Gai
}
