% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp.mean.R
\name{grp.mean}
\alias{grp.mean}
\title{Make a mean dataframe by group from data with concurrent controls}
\usage{
grp.mean(data, group, filter.std = TRUE)
}
\arguments{
\item{data}{original data}

\item{group}{a factor with grouping information}

\item{filter.std}{a logical vevtor, default value is TRUE.}
}
\value{
the grp.mean dataframe
}
\description{
Make a mean dataframe by group from data with concurrent controls
}
