% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.fcm.wss.R
\name{iterate.fcm.wss}
\alias{iterate.fcm.wss}
\title{Fuzzy c-means: choose the clusters number}
\usage{
iterate.fcm.wss(df, m)
}
\arguments{
\item{df}{a dataframe of "grp.mean" class}

\item{m}{a numeric abtained by mestimate function}
}
\value{
a plot
}
\description{
Fuzzy c-means: choose the clusters number
}
\examples{
iterate.fcm.wss(scaledata, mestimate(scaledata))
}
\author{
Zhonghui Gai
}
