% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mestimate.R
\name{mestimate}
\alias{mestimate}
\title{Estimate the fuzzifier step 3}
\usage{
mestimate(df)
}
\arguments{
\item{df}{a dataframe of "grp.mean" class}
}
\value{
a numeric
}
\description{
Estimate the fuzzifier step 3
}
\examples{
m <- mestimate(scaledata)
}
\author{
Zhonghui Gai
}
